Move Project type into Model module
